+++ apps/web/src/app/app.tsx
@@ -1,6 +1,7 @@
 import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 import { HelmetProvider } from 'react-helmet-async';
+import { I18nextProvider } from 'react-i18next';
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
@@ -11,12 +12,14 @@ import { Toaster } from '@eurekabox/ui-kit/components/ui/toaster';
 import { useInitWebCore, useRefreshToken } from '@eurekabox/web-core';
 
 import { Router } from './routes';
+import i18n from '../i18n';
 
 export function App() {
     const queryClient = new QueryClient({
         defaultOptions: {
             queries: {
                 staleTime: Infinity,
+                retry: 1,
             },
         },
     });
@@ -29,20 +32,22 @@ export function App() {
     }
 
     return (
-        <Suspense fallback={LoadingFallback}>
-            <ErrorBoundary FallbackComponent={ErrorFallback}>
-                <HelmetProvider>
-                    <QueryClientProvider client={queryClient}>
-                        <ThemeProvider>
-                            <Router />
-                            <GlobalLoader />
-                            <Toaster />
-                        </ThemeProvider>
-                        {process.env.NODE_ENV !== 'prod' && <ReactQueryDevtools />}
-                    </QueryClientProvider>
-                </HelmetProvider>
-            </ErrorBoundary>
-        </Suspense>
+        <I18nextProvider i18n={i18n}>
+            <Suspense fallback={<LoadingFallback />}>
+                <ErrorBoundary FallbackComponent={ErrorFallback}>
+                    <HelmetProvider>
+                        <QueryClientProvider client={queryClient}>
+                            <ThemeProvider>
+                                <Router />
+                                <GlobalLoader />
+                                <Toaster />
+                            </ThemeProvider>
+                            {process.env.NODE_ENV !== 'prod' && <ReactQueryDevtools />}
+                        </QueryClientProvider>
+                    </HelmetProvider>
+                </ErrorBoundary>
+            </Suspense>
+        </I18nextProvider>
     );
 }