+++ apps/web/src/app/features/editor/routes/index.tsx
@@ -1,0 +1,12 @@
+import { Navigate, Route, Routes } from 'react-router-dom';
+
+import { EditorPage } from '../pages';
+
+export const EditorRoutes = () => {
+    return (
+        <Routes>
+            <Route path="/:contentId" element={<EditorPage />} />
+            <Route path="*" element={<Navigate to="/home" />} />
+        </Routes>
+    );
+};