+++ apps/web/src/app/home/components/dashboard/index.ts
@@ -1,4 +1,0 @@
-export * from './ActionFeed';
-export * from './Overview';
-export * from './RecentDocument';
-export * from './UserStats';