+++ apps/web/src/app/routes/protected/ProtectedRoutes.tsx
@@ -1,9 +1,11 @@
 import { Navigate } from 'react-router-dom';
 
-import { HomeRoutes } from '../../home';
+import { EditorRoutes } from '../../features/editor';
+import { HomeRoutes } from '../../features/home';
 
 export const ProtectedRoutes = [
-    { path: `/home/*`, element: <HomeRoutes /> },
-    { path: '*', element: <Navigate to="/home" replace /> },
     { path: '/', element: <Navigate to="/home" replace /> },
+    { path: `/home`, element: <HomeRoutes /> },
+    { path: `/*`, element: <EditorRoutes /> },
+    { path: '*', element: <Navigate to="/home" replace /> },
 ];