+++ apps/web/src/app/routes/public/PublicRoutes.tsx
@@ -1,6 +1,6 @@
 import { Navigate, useLocation } from 'react-router-dom';
 
-import { AuthRoutes } from '../../auth';
+import { AuthRoutes } from '../../features/auth';
 
 const AuthRedirect = () => {
     const location = useLocation();
@@ -9,6 +9,7 @@ const AuthRedirect = () => {
 
 export const PublicRoutes = [
     { path: `/auth/*`, element: <AuthRoutes /> },
+    { path: '/', element: <AuthRedirect /> },
     {
         path: '*',
         element: <AuthRedirect />,